requirements.txt
streamlit
datetime
yfinance
Prophet
plot_plotly
graph_objs